address maintainer review feedback on PR #1036
- Extract contentSlab into gitindex/slab.go to fix disconnected doc
  comment on indexCatfileBlobs (Go doc spec requires no intervening
  declarations between comment and function)
- Remove GC tuning (GOGC=-1, GOMEMLIMIT=2GiB) — per Go GC guide,
  hard-coded limit risks thrashing for repos with live heap >2GiB and
  violates "don't use when deploying to environments you don't control"
- Add DetermineLanguageIfUnknown in Builder.Add to pre-compute language
  alongside category while content is still available
- Remove redundant DetermineLanguageIfUnknown call from parseSymbols
  (now a no-op since Builder.Add already sets Language)
- Remove DetermineFileCategory fallback loop from sortDocuments (all
  docs come through Builder.Add which already computes Category); fix
  TestFileRank to pre-compute categories on its test documents
- Clean up stale comments referencing removed enry calls and old flow
