+++ cmd/zoekt-git-index/main.go
@@ -21,7 +21,6 @@ import (
 	"log"
 	"os"
 	"path/filepath"
-	"runtime/debug"
 	"runtime/pprof"
 	"strings"
 
@@ -54,16 +53,6 @@ func run() int {
 
 	// Tune GOMAXPROCS to match Linux container CPU quota.
 	_, _ = maxprocs.Set()
-
-	// Optimize GC for batch indexing: disable percentage-based GC and rely
-	// on a memory limit instead. This reduces GC/madvise overhead by ~30%
-	// for large repos. Both can be overridden via GOMEMLIMIT / GOGC env vars.
-	if os.Getenv("GOMEMLIMIT") == "" {
-		debug.SetMemoryLimit(2 << 30) // 2 GiB default
-	}
-	if os.Getenv("GOGC") == "" {
-		debug.SetGCPercent(-1)
-	}
 
 	if *cpuProfile != "" {
 		f, err := os.Create(*cpuProfile)
